requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=0.9.0=pypi_0
arcade=2.4.2=pypi_0
astor=0.8.1=pypi_0
attrs=19.3.0=pypi_0
backcall=0.1.0=pypi_0
blas=1.0=mkl
bleach=3.1.5=pypi_0
ca-certificates=2020.1.1=0
certifi=2020.4.5.2=py37_0
cffi=1.14.3=pypi_0
colorama=0.4.3=pypi_0
comtypes=1.1.7=pypi_0
contextlib2=0.6.0.post1=pypi_0
cudatoolkit=10.0.130=0
cudnn=7.6.4=cuda10.0_0
cycler=0.10.0=pypi_0
cython=0.29.19=pypi_0
d3dshot=0.1.5=pypi_0
decorator=4.4.2=pypi_0
defusedxml=0.6.0=pypi_0
entrypoints=0.3=pypi_0
gast=0.3.3=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.29.0=pypi_0
h5py=2.10.0=pypi_0
icc_rt=2019.0.0=h0cc432a_1
importlib-metadata=1.6.0=pypi_0
intel-openmp=2020.1=216
ipykernel=5.3.0=pypi_0
ipython=7.14.0=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.5.1=pypi_0
jedi=0.17.0=pypi_0
jinja2=2.11.2=pypi_0
joblib=0.15.1=py_0
jsonschema=3.2.0=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=6.1.3=pypi_0
jupyter-console=6.1.0=pypi_0
jupyter-core=4.6.3=pypi_0
keras-applications=1.0.8=pypi_0
keras-preprocessing=1.1.2=pypi_0
keyboard=0.13.5=pypi_0
kiwisolver=1.2.0=pypi_0
libprotobuf=3.11.4=h7bd577a_0
markdown=3.2.2=pypi_0
markupsafe=1.1.1=pypi_0
matplotlib=3.2.1=pypi_0
mistune=0.8.4=pypi_0
mkl=2020.1=216
mkl-service=2.3.0=py37hb782905_0
mkl_fft=1.0.15=py37h14836fe_0
mkl_random=1.1.1=py37h47e9c7a_0
mouseinfo=0.1.3=pypi_0
nbconvert=5.6.1=pypi_0
nbformat=5.0.6=pypi_0
notebook=6.0.3=pypi_0
numpy=1.18.4=pypi_0
numpy-base=1.18.1=py37hc3f5095_1
object-detection=0.1=pypi_0
opencv-python=4.2.0.34=pypi_0
openssl=1.1.1g=he774522_0
packaging=20.4=pypi_0
pandas=1.0.3=pypi_0
pandocfilters=1.4.2=pypi_0
parso=0.7.0=pypi_0
pathfinding=0.0.4=pypi_0
pickleshare=0.7.5=pypi_0
pillow=7.1.2=pypi_0
pip=20.0.2=py37_3
prometheus-client=0.8.0=pypi_0
prompt-toolkit=3.0.5=pypi_0
protobuf=3.12.1=pypi_0
pyautogui=0.9.50=pypi_0
pyclick=0.0.2=pypi_0
pycparser=2.20=pypi_0
pygame=1.9.6=pypi_0
pygetwindow=0.0.8=pypi_0
pyglet=1.5.7=pypi_0
pygments=2.6.1=pypi_0
pymsgbox=1.0.8=pypi_0
pymunk=5.7.0=pypi_0
pyparsing=2.4.7=pypi_0
pyperclip=1.8.0=pypi_0
pyrect=0.1.4=pypi_0
pyrsistent=0.16.0=pypi_0
pyscreeze=0.1.26=pypi_0
python=3.7.7=h81c818b_4
python-dateutil=2.8.1=pypi_0
python3-xlib=0.15=pypi_0
pytiled-parser=0.9.4a3=pypi_0
pytweening=1.0.3=pypi_0
pytz=2020.1=pypi_0
pywin32=227=pypi_0
pywinpty=0.5.7=pypi_0
pyyaml=5.3.1=pypi_0
pyzmq=19.0.1=pypi_0
qtconsole=4.7.4=pypi_0
qtpy=1.9.0=pypi_0
scales=1.0.9=pypi_0
scikit-learn=0.22.1=py37h6288b17_0
scipy=1.4.1=py37h9439919_0
seaborn=0.10.1=pypi_0
send2trash=1.5.0=pypi_0
setuptools=46.4.0=py37_0
six=1.15.0=pypi_0
sqlite=3.31.1=h2a8f88b_1
tensorboard=1.14.0=pypi_0
tensorflow-estimator=1.14.0=pypi_0
tensorflow-gpu=1.14.0=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.8.3=pypi_0
testpath=0.4.4=pypi_0
tornado=6.0.4=pypi_0
traitlets=4.3.3=pypi_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_2
wcwidth=0.1.9=pypi_0
webencodings=0.5.1=pypi_0
werkzeug=1.0.1=pypi_0
wheel=0.34.2=py37_0
widgetsnbextension=3.5.1=pypi_0
wincertstore=0.2=py37_0
wrapt=1.12.1=pypi_0
xlib=0.21=pypi_0
zipp=3.1.0=pypi_0
zlib=1.2.11=h62dcd97_4
